Added linear model
In addition to previous .mod file with different versions of the same
non-linear model, I switched to an exponential utility function for
bonds and money and linearised around a zero-inflation steady state.
Then wrote down the new linear model in dtclin.mod, which simulates it
coherently with previous iterations, adding a flexible and non binding
policy rule.